서버에서만 접근 가능한 객체가 클라이언트로 전달되는 문제 수정
